this one is exactly the same as v3
It uses the open ai tts module and has the goggle tts removed. 
this is just in case i destroy the file in the process of adding the voice selection drop down back into v3. 
see v4 for a back up of the google tts module. where going to have to add the features in v3 to v4 eventually.
Then comes the fun part. adding the opencv crap!
After that we'll have to start over in kotlin or use raspberry pi's for testing!